requirements.txt
discord.py==2.3.2
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio-throttle==1.0.2 